Move mdx-component to basic
