+++ packages/integrations/mdx/test/fixtures/basic/src/components/component/Test.mdx
