+++ packages/integrations/mdx/test/fixtures/basic/src/components/component/WithFragment.mdx
