+++ packages/integrations/mdx/test/fixtures/basic/src/pages/component/glob.astro
@@ -1,6 +1,6 @@
 ---
 import { parse } from 'node:path';
-const components = await Astro.glob('../components/*.mdx');
+const components = await Astro.glob('../../components/component/*.mdx');
 ---
 
 <div data-default-export>
