+++ packages/integrations/mdx/test/fixtures/basic/src/pages/component/index.astro
@@ -1,0 +1,5 @@
+---
+import Test from '../../components/component/Test.mdx';
+---
+
+<Test />
