+++ packages/integrations/mdx/test/fixtures/basic/src/pages/component/w-fragment.astro
@@ -1,0 +1,5 @@
+---
+import WithFragment from '../../components/component/WithFragment.mdx';
+---
+
+<WithFragment />
