--- packages/integrations/mdx/test/fixtures/mdx-component/src/pages/w-fragment.astro
@@ -1,5 +1,0 @@
----
-import WithFragment from '../components/WithFragment.mdx';
----
-
-<WithFragment />
